[cartographer_ros] Add support for 2D SLAM with a depth camera. (#54)
This changes the URDF and timeout for the TurtleBot configuration,
and adds a Dockerfile with everything that is needed to run 2D SLAM
with a depth camera. Verified to work on a TurtleBot 2 with an
ASUS Xtion.

And tiny style fixes.

Original commit:
googlecartographer/cartographer_ros@0e26c9feac33cc8a86159aeb2ac270d312a9670c