[cartographer_ros] Increase 3D global localization score threshold. (#587)
With googlecartographer/cartographer#640 submitted,
global localization in 3D seems to behave as expected,
so there is no longer a need for a low minimum score.

We increase the minimum score to be slightly above the
minimum score used in the same trajectory. This is to
reduce the amount of outliers which are worse in this
case. This change also moves it into the backpack_3d.lua
file since it should also apply to multi-trajectory SLAM.

Original commit:
googlecartographer/cartographer_ros@7540dfa67491fba37ceb134c6363f2539c9645e2