[cartographer] Make LocalTrajectoryBuilder optional. (#799)
This is to prepare for the cloud-based mapping case where the robot is running cartographer in pure-localization mode and the cloud instance only solves the global problem. In that case a `LocalTrajectoryBuilder` need to be instantiated in for the cloud instance.

Original commit:
googlecartographer/cartographer@8c7c4e3d2ad74c892cc77a3b54757bab411b90cb