[cartographer_ros] Removing unless from argument (#994)
The roslaunch API throws an exception if setting the argument but not using it. 
This now forces any including launch files to provide an argument for it, even if it might not be used in the `no_rviz` case.

Original commit:
googlecartographer/cartographer_ros@744c03a31fded1e115768c5667d226e5bf68f9a1