[cartographer_ros] Add explicit --save_state_filename to offline node options. (#1204)
This allows to serialize the state also when no bagfile is given, e.g.
when the offline node is used to run an optimization and/or trimming
configuration on a pbstream. Or simply when one wants to use a custom
name directly. This doesn't break compatibility with the previous CLI.

Original commit:
googlecartographer/cartographer_ros@9f8bf5b24537e60858df395a30fc8fe06264ad50